Cleaned up code, added docstring, implemented dark-mode toggle
